Add help subcommand to all commands
Related to issue #4 in github, I've created a
small improvement that allows users to use
"@Dave ping help" to get the help info for
the ping command instead of forcing the use
of "@Dave help" to display the help text of
all commands. "@Dave help" now only returns
a list of all commands.